chore(channel-web): refactoring
