+++ modules/channel-web/package.json
@@ -34,7 +34,7 @@
     "query-string": "^5.0.1",
     "uuid": "^3.1.0",
     "react-linkify": "^0.2.2",
-    "react-slick": "^0.23.2",
+    "react-slick": "^0.18.0",
     "slick-carousel": "^1.8.1",
     "snarkdown": "^1.2.2"
   }
