+++ modules/channel-web/src/views/components/BotInfo.scss
