+++ modules/channel-web/src/views/components/index.jsx
@@ -1,0 +1,9 @@
+export { QuickReplies } from './messages/renderer/QuickReplies'
+export { FileMessage } from './messages/renderer/FileMessage'
+export { FileInput } from './messages/renderer/FileInput'
+export { Carousel } from './messages/renderer/Carousel'
+export { LoginPrompt } from './messages/renderer/LoginPrompt'
+export { Text } from './messages/renderer/Text'
+export { Form } from './messages/renderer/Form'
+export { MessageList } from './messages/MessageList'
+export { MessageGroup } from './messages/MessageGroup'
