+++ modules/channel-web/src/views/components/style.scss
@@ -1,0 +1,217 @@
+.avatar {
+  margin-top: 0.15rem;
+  margin-right: 0.75rem;
+  -ms-flex-negative: 0;
+  flex-shrink: 0;
+  width: 1.25rem;
+  height: 1.25rem;
+
+  .color {
+    fill: currentColor;
+  }
+
+  .picture {
+    border-radius: 50%;
+    background-size: cover;
+    background-position: 50%;
+    width: 100%;
+    height: 100%;
+  }
+}
+
+.otherFile {
+  padding: 0.5rem 0.75rem;
+}
+
+.formOverlay {
+  background-color: rgba(0, 0, 0, 0.5);
+  width: 100vw;
+  position: absolute;
+  top: 0;
+  z-index: 1000000;
+  height: 100vh;
+  left: 0;
+  padding: 5%;
+}
+
+.formContainer {
+  width: 320px;
+  margin: auto;
+  background-color: #fff;
+  padding: 25px;
+  border-radius: 5px;
+
+  .formClose {
+    display: inline-block;
+    float: right;
+    vertical-align: top;
+    position: relative;
+    top: 0;
+    cursor: pointer;
+  }
+
+  .formTextarea {
+    display: block;
+    width: 100%;
+    height: 80px;
+    min-height: 80px;
+    max-width: 270px;
+    padding: 6px 12px;
+    font-size: 14px;
+    line-height: 1.42857143;
+    color: #555;
+    background-color: #fff;
+    background-image: none;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    -webkit-box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
+    box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
+    -webkit-transition: border-color ease-in-out 0.15s, -webkit-box-shadow ease-in-out 0.15s;
+    -o-transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
+    transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
+  }
+  .formInput {
+    display: block;
+    width: 100%;
+    height: 34px;
+    padding: 6px 12px;
+    font-size: 14px;
+    line-height: 1.42857143;
+    color: #555;
+    background-color: #fff;
+    background-image: none;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    -webkit-box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
+    box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
+    -webkit-transition: border-color ease-in-out 0.15s, -webkit-box-shadow ease-in-out 0.15s;
+    -o-transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
+    transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
+  }
+  .formTitle {
+    color: #000;
+    font-size: 18px;
+  }
+  label {
+    display: inline-block;
+    max-width: 100%;
+    margin-bottom: 5px;
+    font-weight: bold;
+    font-size: 14px;
+  }
+  .formGroup {
+    margin-bottom: 15px;
+    margin-top: 10px;
+  }
+  .buttonLayer {
+    text-align: right;
+    .formSubmit {
+      display: inline-block;
+      padding: 6px 12px;
+      margin-bottom: 0;
+      font-size: 14px;
+      font-weight: normal;
+      line-height: 1.42857143;
+      text-align: center;
+      white-space: nowrap;
+      vertical-align: middle;
+      -ms-touch-action: manipulation;
+      touch-action: manipulation;
+      cursor: pointer;
+      -webkit-user-select: none;
+      -moz-user-select: none;
+      -ms-user-select: none;
+      user-select: none;
+      background-image: none;
+      border-radius: 4px;
+      color: #333;
+      background-color: #fff;
+      border: 1px solid #ccc;
+    }
+    .formSubmit:hover {
+      background-color: #ccc;
+    }
+  }
+}
+
+.loginPromptContainer {
+  label {
+    display: block;
+  }
+
+  .loginInput {
+    height: 30px;
+    border: none;
+    outline: none;
+    background-color: #f8f8f8;
+    border-bottom: 1px solid #bbb;
+    width: 200px;
+    text-align: center;
+    margin: 10px;
+    font-size: 14px;
+  }
+
+  .loginButton {
+    display: block;
+    font-size: 16px;
+    height: 30px;
+    width: 130px;
+    text-align: center;
+    margin: auto;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    border: none;
+    outline: none;
+    background-color: black;
+    color: white;
+  }
+}
+
+@keyframes typingAnimation {
+  0% {
+    transform: translateY(0px);
+  }
+  28% {
+    transform: translateY(-6px);
+  }
+  44% {
+    transform: translateY(0px);
+  }
+}
+
+.typingGroup {
+  align-items: center;
+  display: flex;
+  height: 17px;
+
+  .typingBubble {
+    animation: typingAnimation 1s infinite ease-in-out;
+    border-radius: 50%;
+    height: 6px;
+    margin-right: 4px;
+    vertical-align: middle;
+    width: 6px;
+    white-space: pre-wrap;
+    font-size: 16px;
+    direction: ltr;
+
+    &:nth-child(1) {
+      animation-delay: 200ms;
+    }
+    &:nth-child(2) {
+      animation-delay: 300ms;
+    }
+    &:nth-child(3) {
+      animation-delay: 400ms;
+    }
+  }
+}
+
+/* IE10+ CSS styles go here */
+@media all and (-ms-high-contrast: none), (-ms-high-contrast: active) {
+  .typingGroup {
+    .typingBubble {
+      animation: initial;
+    }
+  }
+}
