--- modules/channel-web/src/views/lite/async-debounce.js
@@ -1,17 +1,0 @@
-const asyncDebounce = async timeMs => {
-  let lastClickInMs = undefined
-
-  const debounce = promise => {
-    const now = Date.now()
-    if (!lastClickInMs) {
-      lastClickInMs = now
-    }
-
-    if (now - lastClickInMs > timeMs) {
-      lastClickInMs = now
-      return promise
-    }
-  }
-}
-
-export default asyncDebounce
