+++ modules/channel-web/src/views/lite/main.jsx
@@ -12,7 +12,6 @@ import Convo from './convo'
 import Side from './side'
 
 import style from './style.scss'
-import asyncDebounce from './async-debounce'
 
 const _values = obj => Object.keys(obj).map(x => obj[x])
 
@@ -56,8 +55,6 @@ export default class Web extends React.Component {
       ReactGA.initialize('UA-90044826-2')
       ReactGA.event({ category: 'WebChat', action: 'render', nonInteraction: true })
     }
-
-    this.quickReplyDebounce = asyncDebounce(1000)
 
     const { options } = queryString.parse(location.search)
     const { config } = JSON.parse(decodeURIComponent(options || '{}'))
@@ -159,12 +156,6 @@ export default class Web extends React.Component {
     this.setUserId().then(this.fetchData)
   }
 
-  sendUserEvent = data => {
-    const userId = window.__BP_VISITOR_ID
-    const url = `/mod/channel-web/events/${userId}`
-    return this.props.bp.axios.post(url, data, this.axiosConfig)
-  }
-
   handleIframeApi = ({ data: { action, payload } }) => {
     if (action === 'configure') {
       if (payload.userId) {
@@ -181,7 +172,7 @@ export default class Web extends React.Component {
         this.setState({ textToSend: text })
         this.handleSendMessage()
       } else {
-        return this.sendUserEvent({ type, payload })
+        return this.handleSendData({ type, payload })
       }
     }
   }
@@ -483,40 +474,7 @@ export default class Web extends React.Component {
     })
   }
 
-  handleAddEmoji = emoji => {
-    this.setState({
-      textToSend: this.state.textToSend + emoji.native + ' '
-    })
-  }
-
-  handleSendQuickReply = (title, payload) => {
-    const promise = this.handleSendData({
-      type: 'quick_reply',
-      text: title,
-      data: { payload }
-    })
-
-    return this.quickReplyDebounce.debounce(promise)
-  }
-
-  handleSendForm = (fields, formId, repr) => {
-    return this.handleSendData({
-      type: 'form',
-      formId: formId,
-      text: repr,
-      data: fields
-    })
-  }
-
-  handleLoginPrompt = (username, password) => {
-    return this.handleSendData({
-      type: 'login_prompt',
-      text: 'Provided login information',
-      data: { username, password }
-    })
-  }
-
-  handleFileUploadSend = (title, payload, file) => {
+  handleFileUpload = (title, payload, file) => {
     const userId = window.__BP_VISITOR_ID
     const url = `/mod/channel-web/messages/${userId}/files`
     const config = { params: { conversationId: this.state.currentConversationId }, ...this.axiosConfig }
@@ -529,6 +487,13 @@ export default class Web extends React.Component {
 
   handleSendData = data => {
     const userId = window.__BP_VISITOR_ID
+    const msgTypes = ['text', 'quick_reply', 'form', 'login_prompt', 'visit', 'postback']
+
+    if (!msgTypes.includes(data.type)) {
+      const url = `/mod/channel-web/events/${userId}`
+      return this.props.bp.axios.post(url, data, this.axiosConfig)
+    }
+
     const url = `/mod/channel-web/messages/${userId}`
     const config = { params: { conversationId: this.state.currentConversationId }, ...this.axiosConfig }
 
@@ -591,9 +556,8 @@ export default class Web extends React.Component {
     }
     return (
       <button
-        className={classnames('bp-widget-btn', style[this.state.widgetTransition], style.floatingButton)}
+        className={classnames('bpw-widget-btn', style[this.state.widgetTransition], style.floatingButton)}
         onClick={this.handleButtonClicked}
-        style={{ backgroundColor: this.state.config.foregroundColor }}
       >
         <i>{this.state.view === 'convo' ? this.renderCloseIcon() : this.renderOpenIcon()}</i>
         {this.state.unreadCount > 0 ? this.renderUncountMessages() : null}
@@ -654,10 +618,6 @@ export default class Web extends React.Component {
     this.downaloadFile(file.name, blobFile)
   }
 
-  startConversation = cb => {
-    return this.sendUserEvent({ payload: { type: 'request_start_conversation' } }).then(cb)
-  }
-
   renderSide() {
     return (
       <Side
@@ -669,21 +629,16 @@ export default class Web extends React.Component {
         unreadCount={this.state.unreadCount}
         currentConversation={this.state.currentConversation}
         conversations={this.state.conversations}
-        addEmojiToText={this.handleAddEmoji}
         onClose={!this.props.fullscreen ? this.handleClosePanel : null}
         onResetSession={this.handleSessionReset}
         onSwitchConvo={this.handleSwitchConvo}
         onTextSend={this.handleSendMessage}
         recallHistory={this.handleRecallHistory}
         onTextChanged={this.handleTextChanged}
-        onQuickReplySend={this.handleSendQuickReply}
-        onFormSend={this.handleSendForm}
-        onFileUploadSend={this.handleFileUploadSend}
-        onLoginPromptSend={this.handleLoginPrompt}
+        onFileUpload={this.handleFileUpload}
         onSendData={this.handleSendData}
         downloadConversation={this.downloadConversation}
         createConversation={this.createConversation}
-        startConversation={this.startConversation}
         botInfo={this.state.botInfo}
       />
     )
@@ -702,6 +657,7 @@ export default class Web extends React.Component {
 
     return (
       <div className={classnames('bp-chat', style.web)} onFocus={this.handleResetUnreadCount}>
+        <link rel="stylesheet" type="text/css" href={'/assets/modules/channel-web/default.css'} />
         {stylesheet && stylesheet.length && <link rel="stylesheet" type="text/css" href={stylesheet} />}
         {view}
       </div>
