--- modules/channel-web/src/views/lite/messages/misc.js
@@ -1,15 +1,0 @@
-export function hexToRGBA(hex, alpha = 1) {
-  const r = parseInt(hex.slice(1, 3), 16)
-  const g = parseInt(hex.slice(3, 5), 16)
-  const b = parseInt(hex.slice(5, 7), 16)
-  return `rgba(${r}, ${g}, ${b}, ${alpha})`
-}
-
-export const getOverridedComponent = (overrides, componentName) => {
-  if (overrides && overrides[componentName]) {
-    const { module, component } = overrides[componentName]
-    if (module && component) {
-      return window.botpress[module][component]
-    }
-  }
-}
