+++ modules/channel-web/src/views/lite/side/style.scss
@@ -308,7 +308,7 @@
       overflow-x: hidden;
 
       .item {
-        margin-left: 15px;
+        padding-left: 15px;
         height: 4.7rem;
         border-bottom: 1px solid #eee;
         cursor: pointer;
