+++ modules/channel-web/src/views/lite/tools.js
@@ -1,0 +1,24 @@
+export const getOverridedComponent = (overrides, componentName) => {
+  if (overrides && overrides[componentName]) {
+    const { module, component } = overrides[componentName]
+    if (module && component) {
+      return window.botpress[module][component]
+    }
+  }
+}
+
+export const asyncDebounce = async timeMs => {
+  let lastClickInMs = undefined
+
+  const debounce = promise => {
+    const now = Date.now()
+    if (!lastClickInMs) {
+      lastClickInMs = now
+    }
+
+    if (now - lastClickInMs > timeMs) {
+      lastClickInMs = now
+      return promise
+    }
+  }
+}
